build(core/io): move headers to the right place
after smp had been moved to io crate
