+++ core/embed/io/nrf/rust_smp.h
