+++ core/embed/io/nrf/stm32u5/nrf_uart.c
@@ -26,7 +26,7 @@
 #include <sys/mpu.h>
 
 #include "../nrf_internal.h"
-#include "rust_smp.h"
+#include "../rust_smp.h"
 #include "sys/systick.h"
 
 extern nrf_driver_t g_nrf_driver;
